+++ test/legacy/CollateralErc20.js
@@ -1,60 +1,71 @@
 'use strict';
 
-const { contract, web3 } = require('hardhat');
+const { artifacts, contract, web3 } = require('hardhat');
 
-const { assert, addSnapshotBeforeRestoreAfterEach } = require('./common');
+const { assert, addSnapshotBeforeRestoreAfterEach } = require('../contracts/common');
 
 const BN = require('bn.js');
 
-const { fastForward, getEthBalance, toUnit, fromUnit } = require('../utils')();
+const PublicEST8Decimals = artifacts.require('PublicEST8Decimals');
 
-const { setupAllContracts } = require('./setup');
+const { fastForward, toUnit } = require('../utils')();
+
+const { setupAllContracts, setupContract } = require('../contracts/setup');
 
 const {
 	ensureOnlyExpectedMutativeFunctions,
 	setStatus,
 	setupPriceAggregators,
 	updateAggregatorRates,
-} = require('./helpers');
+} = require('../contracts/helpers');
 
-const { toBytes32 } = require('../..');
+const {
+	toBytes32,
+	constants: { ZERO_ADDRESS },
+} = require('../..');
 
-contract('CollateralEth', async accounts => {
-	const YEAR = 31556926;
+let CollateralManager;
+let CollateralManagerState;
+let ProxyERC20;
+let TokenState;
+
+contract('CollateralErc20', async accounts => {
+	const YEAR = 31536000;
 	const INTERACTION_DELAY = 300;
 
 	const sUSD = toBytes32('sUSD');
 	const sETH = toBytes32('sETH');
 	const sBTC = toBytes32('sBTC');
 
-	const oneETH = toUnit(1);
-	const twoETH = toUnit(2);
-	const fiveETH = toUnit(5);
-	const tenETH = toUnit(10);
-	const twentyETH = toUnit(20);
+	const oneRenBTC = web3.utils.toBN('100000000');
+	const twoRenBTC = web3.utils.toBN('200000000');
+	const fiveRenBTC = web3.utils.toBN('500000000');
 
 	const onesUSD = toUnit(1);
 	const tensUSD = toUnit(10);
 	const oneHundredsUSD = toUnit(100);
-	const fiveHundredsUSD = toUnit(500);
+	const oneThousandsUSD = toUnit(1000);
+	const fiveThousandsUSD = toUnit(5000);
 
 	let tx;
 	let loan;
 	let id;
+	let proxy, tokenState;
 
-	const [, owner, , account1, account2] = accounts;
+	const [deployerAccount, owner, , , account1, account2] = accounts;
 
-	let ceth,
+	let cerc20,
 		managerState,
-		manager,
-		issuer,
-		synths,
 		feePool,
 		exchangeRates,
 		addressResolver,
 		sUSDSynth,
-		sETHSynth,
+		sBTCSynth,
+		renBTC,
 		systemStatus,
+		synths,
+		manager,
+		issuer,
 		debtCache,
 		FEE_ADDRESS;
 
@@ -70,80 +81,12 @@ contract('CollateralEth', async accounts => {
 		});
 	};
 
-	const issuesETHToAccount = async (issueAmount, receiver) => {
-		await sETHSynth.issue(receiver, issueAmount, { from: owner });
+	const issuesBTCtoAccount = async (issueAmount, receiver) => {
+		await sBTCSynth.issue(receiver, issueAmount, { from: owner });
 	};
 
-	const setupMultiCollateral = async () => {
-		synths = ['sUSD', 'sETH'];
-		({
-			SystemStatus: systemStatus,
-			ExchangeRates: exchangeRates,
-			SynthsUSD: sUSDSynth,
-			SynthsETH: sETHSynth,
-			FeePool: feePool,
-			AddressResolver: addressResolver,
-			Issuer: issuer,
-			DebtCache: debtCache,
-			CollateralEth: ceth,
-			CollateralManager: manager,
-			CollateralManagerState: managerState,
-		} = await setupAllContracts({
-			accounts,
-			synths,
-			contracts: [
-				'Synthetix',
-				'FeePool',
-				'AddressResolver',
-				'ExchangeRates',
-				'Exchanger',
-				'SystemStatus',
-				'Issuer',
-				'CollateralEth',
-				'CollateralUtil',
-				'CollateralManager',
-				'CollateralManagerState',
-				'DebtCache',
-			],
-		}));
-
-		await setupPriceAggregators(exchangeRates, owner, [sBTC, sETH]);
-
-		await managerState.setAssociatedContract(manager.address, { from: owner });
-
-		FEE_ADDRESS = await feePool.FEE_ADDRESS();
-
-		await addressResolver.importAddresses(
-			[toBytes32('CollateralEth'), toBytes32('CollateralManager')],
-			[ceth.address, manager.address],
-			{
-				from: owner,
-			}
-		);
-
-		await ceth.rebuildCache();
-		await manager.rebuildCache();
-		await debtCache.rebuildCache();
-		await feePool.rebuildCache();
-		await issuer.rebuildCache();
-
-		await manager.addCollaterals([ceth.address], { from: owner });
-
-		await ceth.addSynths(
-			['SynthsUSD', 'SynthsETH'].map(toBytes32),
-			['sUSD', 'sETH'].map(toBytes32),
-			{ from: owner }
-		);
-
-		await manager.addSynths(
-			['SynthsUSD', 'SynthsETH'].map(toBytes32),
-			['sUSD', 'sETH'].map(toBytes32),
-			{ from: owner }
-		);
-		// rebuild the cache to add the synths we need.
-		await manager.rebuildCache();
-
-		await ceth.setIssueFeeRate(toUnit('0.001'), { from: owner });
+	const issueRenBTCtoAccount = async (issueAmount, receiver) => {
+		await renBTC.transfer(receiver, issueAmount, { from: owner });
 	};
 
 	const updateRatesWithDefaults = async () => {
@@ -155,7 +98,148 @@ contract('CollateralEth', async accounts => {
 		await updateRatesWithDefaults();
 	};
 
+	const deployCollateral = async ({
+		owner,
+		manager,
+		resolver,
+		collatKey,
+		minColat,
+		minSize,
+		underCon,
+		decimals,
+	}) => {
+		return setupContract({
+			accounts,
+			contract: 'CollateralErc20',
+			args: [owner, manager, resolver, collatKey, minColat, minSize, underCon, decimals],
+		});
+	};
+
+	const setupMultiCollateral = async () => {
+		synths = ['sUSD', 'sBTC'];
+		({
+			SystemStatus: systemStatus,
+			ExchangeRates: exchangeRates,
+			SynthsUSD: sUSDSynth,
+			SynthsBTC: sBTCSynth,
+			FeePool: feePool,
+			AddressResolver: addressResolver,
+			Issuer: issuer,
+			DebtCache: debtCache,
+		} = await setupAllContracts({
+			accounts,
+			synths,
+			contracts: [
+				'Synthetix',
+				'FeePool',
+				'AddressResolver',
+				'ExchangeRates',
+				'SystemStatus',
+				'Issuer',
+				'DebtCache',
+				'Exchanger',
+				'CollateralUtil',
+			],
+		}));
+
+		await setupPriceAggregators(exchangeRates, owner, [sBTC, sETH]);
+
+		managerState = await CollateralManagerState.new(owner, ZERO_ADDRESS, { from: deployerAccount });
+
+		const maxDebt = toUnit(10000000);
+
+		manager = await CollateralManager.new(
+			managerState.address,
+			owner,
+			addressResolver.address,
+			maxDebt,
+			0,
+			0,
+			0,
+			{
+				from: deployerAccount,
+			}
+		);
+
+		await managerState.setAssociatedContract(manager.address, { from: owner });
+
+		FEE_ADDRESS = await feePool.FEE_ADDRESS();
+
+		// the owner is the associated contract, so we can simulate
+		proxy = await ProxyERC20.new(owner, {
+			from: deployerAccount,
+		});
+		tokenState = await TokenState.new(owner, ZERO_ADDRESS, { from: deployerAccount });
+
+		renBTC = await PublicEST8Decimals.new(
+			proxy.address,
+			tokenState.address,
+			'Some Token',
+			'TOKEN',
+			toUnit('1000'),
+			owner,
+			{
+				from: deployerAccount,
+			}
+		);
+
+		await tokenState.setAssociatedContract(owner, { from: owner });
+		await tokenState.setBalanceOf(owner, toUnit('1000'), { from: owner });
+		await tokenState.setAssociatedContract(renBTC.address, { from: owner });
+
+		await proxy.setTarget(renBTC.address, { from: owner });
+
+		cerc20 = await deployCollateral({
+			owner: owner,
+			manager: manager.address,
+			resolver: addressResolver.address,
+			collatKey: sBTC,
+			minColat: toUnit(1.5),
+			minSize: toUnit(0.1),
+			underCon: renBTC.address,
+			decimals: 8,
+		});
+
+		await addressResolver.importAddresses(
+			[toBytes32('CollateralErc20'), toBytes32('CollateralManager')],
+			[cerc20.address, manager.address],
+			{
+				from: owner,
+			}
+		);
+
+		await feePool.rebuildCache();
+		await manager.rebuildCache();
+		await issuer.rebuildCache();
+		await debtCache.rebuildCache();
+
+		await manager.addCollaterals([cerc20.address], { from: owner });
+
+		await cerc20.addSynths(
+			['SynthsUSD', 'SynthsBTC'].map(toBytes32),
+			['sUSD', 'sBTC'].map(toBytes32),
+			{ from: owner }
+		);
+
+		await manager.addSynths(
+			['SynthsUSD', 'SynthsBTC'].map(toBytes32),
+			['sUSD', 'sBTC'].map(toBytes32),
+			{ from: owner }
+		);
+		// rebuild the cache to add the synths we need.
+		await manager.rebuildCache();
+
+		// Issue ren and set allowance
+		await issueRenBTCtoAccount(100 * 1e8, account1);
+		await renBTC.approve(cerc20.address, 100 * 1e8, { from: account1 });
+	};
+
 	before(async () => {
+		CollateralManager = artifacts.require(`CollateralManager`);
+		CollateralManagerState = artifacts.require('CollateralManagerState');
+		ProxyERC20 = artifacts.require(`ProxyERC20`);
+		TokenState = artifacts.require(`TokenState`);
+
 		await setupMultiCollateral();
 	});
 
@@ -165,27 +249,28 @@ contract('CollateralEth', async accounts => {
 		await updateRatesWithDefaults();
 
 		await issuesUSDToAccount(toUnit(1000), owner);
-		await issuesETHToAccount(toUnit(10), owner);
+		await issuesBTCtoAccount(toUnit(10), owner);
 
 		await debtCache.takeDebtSnapshot();
 	});
 
 	it('should set constructor params on deployment', async () => {
-		// assert.equal(await ceth.proxy(), account1);
-		assert.equal(await ceth.owner(), owner);
-		assert.equal(await ceth.resolver(), addressResolver.address);
-		assert.equal(await ceth.collateralKey(), sETH);
-		assert.equal(await ceth.synths(0), toBytes32('SynthsUSD'));
-		assert.equal(await ceth.synths(1), toBytes32('SynthsETH'));
-		assert.bnEqual(await ceth.minCratio(), toUnit('1.3'));
-		assert.bnEqual(await ceth.minCollateral(), toUnit('2'));
+		assert.equal(await cerc20.owner(), owner);
+		assert.equal(await cerc20.resolver(), addressResolver.address);
+		assert.equal(await cerc20.collateralKey(), sBTC);
+		assert.equal(await cerc20.synths(0), toBytes32('SynthsUSD'));
+		assert.equal(await cerc20.synths(1), toBytes32('SynthsBTC'));
+		assert.bnEqual(await cerc20.minCratio(), toUnit(1.5));
+		assert.bnEqual(await cerc20.minCollateral(), toUnit(0.1));
+		assert.equal(await cerc20.underlyingContract(), renBTC.address);
+		assert.bnEqual(await cerc20.underlyingContractDecimals(), await renBTC.decimals());
 	});
 
 	it('should ensure only expected functions are mutative', async () => {
 		ensureOnlyExpectedMutativeFunctions({
-			abi: ceth.abi,
+			abi: cerc20.abi,
 			ignoreParents: ['Owned', 'Pausable', 'MixinResolver', 'Proxy', 'Collateral'],
-			expected: ['open', 'close', 'deposit', 'repay', 'withdraw', 'liquidate', 'claim', 'draw'],
+			expected: ['open', 'close', 'deposit', 'repay', 'withdraw', 'liquidate', 'draw'],
 		});
 	});
 
@@ -199,59 +284,59 @@ contract('CollateralEth', async accounts => {
 	});
 
 	// PUBLIC VIEW TESTS
-
 	describe('cratio test', async () => {
 		describe('sUSD loans', async () => {
 			beforeEach(async () => {
-				tx = await ceth.open(oneHundredsUSD, sUSD, {
-					value: twoETH,
+				tx = await cerc20.open(oneRenBTC, fiveThousandsUSD, sUSD, {
 					from: account1,
 				});
 
 				id = getid(tx);
+				loan = await cerc20.loans(id);
 			});
 
 			it('when we issue at 200%, our c ratio is 200%', async () => {
-				const ratio = await ceth.collateralRatio(id);
+				const ratio = await cerc20.collateralRatio(id);
 				assert.bnEqual(ratio, toUnit(2));
 			});
 
 			it('when the price falls by 25% our c ratio is 150%', async () => {
-				await updateAggregatorRates(exchangeRates, null, [sETH], [toUnit(75)]);
-				const ratio = await ceth.collateralRatio(id);
+				await updateAggregatorRates(exchangeRates, null, [sBTC], [7500].map(toUnit));
+
+				const ratio = await cerc20.collateralRatio(id);
 				assert.bnEqual(ratio, toUnit(1.5));
 			});
 
 			it('when the price increases by 100% our c ratio is 400%', async () => {
-				await updateAggregatorRates(exchangeRates, null, [sETH], [toUnit(200)]);
-				const ratio = await ceth.collateralRatio(id);
+				await updateAggregatorRates(exchangeRates, null, [sBTC], [20000].map(toUnit));
+				const ratio = await cerc20.collateralRatio(id);
 				assert.bnEqual(ratio, toUnit(4));
 			});
 
-			it('when the price falls by 50% our cratio is 100%', async () => {
-				await updateAggregatorRates(exchangeRates, null, [sETH], [toUnit(50)]);
-				const ratio = await ceth.collateralRatio(id);
+			it('when the price fallsby 50% our cratio is 100%', async () => {
+				await updateAggregatorRates(exchangeRates, null, [sBTC], [5000].map(toUnit));
+				const ratio = await cerc20.collateralRatio(id);
 				assert.bnEqual(ratio, toUnit(1));
 			});
 		});
-		describe('sETH loans', async () => {
+		describe('sBTC loans', async () => {
 			beforeEach(async () => {
-				tx = await ceth.open(oneETH, sETH, {
-					value: twoETH,
+				tx = await cerc20.open(twoRenBTC, toUnit(1), sBTC, {
 					from: account1,
 				});
 
 				id = getid(tx);
+				loan = await cerc20.loans(id);
 			});
 
 			it('when we issue at 200%, our c ratio is 200%', async () => {
-				const ratio = await ceth.collateralRatio(id);
+				const ratio = await cerc20.collateralRatio(id);
 				assert.bnEqual(ratio, toUnit(2));
 			});
 
 			it('price changes should not change the cratio', async () => {
-				await updateAggregatorRates(exchangeRates, null, [sETH], [toUnit(75)]);
-				const ratio = await ceth.collateralRatio(id);
+				await updateAggregatorRates(exchangeRates, null, [sBTC], [75].map(toUnit));
+				const ratio = await cerc20.collateralRatio(id);
 				assert.bnEqual(ratio, toUnit(2));
 			});
 		});
@@ -259,19 +344,55 @@ contract('CollateralEth', async accounts => {
 
 	describe('max loan test', async () => {
 		it('should convert correctly', async () => {
-			// $270 worth of eth should allow 200 sUSD to be issued.
-			const sUSDAmount = await ceth.maxLoan(toUnit('2.6'), sUSD);
+			// $150 worth of btc should allow 100 sUSD to be issued.
+			const sUSDAmount = await cerc20.maxLoan(toUnit(0.015), sUSD);
 
-			assert.bnClose(sUSDAmount, toUnit('200'), '100');
+			assert.bnClose(sUSDAmount, toUnit(100), 100);
 
-			// $270 worth of eth should allow $200 (0.02) of sBTC to be issued.
-			const sBTCAmount = await ceth.maxLoan(toUnit('2.6'), sBTC);
+			// $150 worth of btc should allow $100 (1) of sETH to be issued.
+			const sETHAmount = await cerc20.maxLoan(toUnit(0.015), sETH);
 
-			assert.bnEqual(sBTCAmount, toUnit('0.02'));
+			assert.bnEqual(sETHAmount, toUnit(1));
 		});
 	});
 
-	// END PUBLIC VIEW TESTS
+	describe('scaling collateral test', async () => {
+		it('should scale up 1e8 to 1e18 correctly', async () => {
+			// Scaling up 1 renBTC to 18 decimals works.
+			const scaledCollateral = await cerc20.scaleUpCollateral(oneRenBTC);
+
+			assert.bnEqual(scaledCollateral, toUnit(1));
+		});
+
+		it('should scaled up 1.23456789 correctly', async () => {
+			// Scaling up 1.2345678 renBTC to 8 decimals works.
+			const bal = 123456789;
+			const scaledCollateral = await cerc20.scaleUpCollateral(bal);
+
+			assert.bnEqual(scaledCollateral, toUnit('1.23456789'));
+		});
+
+		it('should scale down 1e18 to 1e8 correctly', async () => {
+			// Scaling down 1.2345678 renBTC to 8 decimals works.
+			const scaledCollateral = await cerc20.scaleDownCollateral(toUnit('1'));
+
+			assert.bnEqual(scaledCollateral, oneRenBTC);
+		});
+
+		it('should scale down 1.23456789 correctly', async () => {
+			// Scaling down 1 renBTC to 8 decimals works.
+			const scaledCollateral = await cerc20.scaleDownCollateral(toUnit('1.23456789'));
+
+			assert.bnEqual(scaledCollateral, 123456789);
+		});
+
+		it('if more than 8 decimals come back, it truncates and does not round', async () => {
+			// If we round, we might run out of ren in the contract.
+			const scaledCollateral = await cerc20.scaleDownCollateral(toUnit('1.23456789999999999'));
+
+			assert.bnEqual(scaledCollateral, 123456789);
+		});
+	});
 
 	// LOAN INTERACTIONS
 
@@ -284,7 +405,7 @@ contract('CollateralEth', async accounts => {
 					});
 					it('then calling openLoan() reverts', async () => {
 						await assert.revert(
-							ceth.open(onesUSD, sUSD, { value: twoETH, from: account1 }),
+							cerc20.open(oneRenBTC, onesUSD, sUSD, { from: account1 }),
 							'Operation prohibited'
 						);
 					});
@@ -293,8 +414,7 @@ contract('CollateralEth', async accounts => {
 							await setStatus({ owner, systemStatus, section, suspend: false });
 						});
 						it('then calling openLoan() succeeds', async () => {
-							await ceth.open(onesUSD, sUSD, {
-								value: twoETH,
+							await cerc20.open(oneRenBTC, onesUSD, sUSD, {
 								from: account1,
 							});
 						});
@@ -307,16 +427,16 @@ contract('CollateralEth', async accounts => {
 				});
 				it('then calling openLoan() reverts', async () => {
 					await assert.revert(
-						ceth.open(onesUSD, sUSD, { value: twoETH, from: account1 }),
+						cerc20.open(oneRenBTC, onesUSD, sUSD, { from: account1 }),
 						'Invalid rate'
 					);
 				});
-				describe('when ETH gets a rate', () => {
+				describe('when BTC gets a rate', () => {
 					beforeEach(async () => {
 						await updateRatesWithDefaults();
 					});
 					it('then calling openLoan() succeeds', async () => {
-						await ceth.open(onesUSD, sUSD, { value: twoETH, from: account1 });
+						await cerc20.open(oneRenBTC, onesUSD, sUSD, { from: account1 });
 					});
 				});
 			});
@@ -325,48 +445,56 @@ contract('CollateralEth', async accounts => {
 		describe('revert conditions', async () => {
 			it('should revert if they request a currency that is not supported', async () => {
 				await assert.revert(
-					ceth.open(onesUSD, toBytes32('sJPY'), { value: twoETH, from: account1 }),
+					cerc20.open(oneRenBTC, onesUSD, toBytes32('sJPY'), { from: account1 }),
 					'Not allowed to issue'
 				);
 			});
 
 			it('should revert if they send 0 collateral', async () => {
 				await assert.revert(
-					ceth.open(onesUSD, sUSD, { value: oneETH, from: account1 }),
+					cerc20.open(toUnit(0), onesUSD, sUSD, { from: account1 }),
 					'Not enough collateral'
 				);
 			});
 
 			it('should revert if the requested loan exceeds borrowing power', async () => {
 				await assert.revert(
-					ceth.open(fiveHundredsUSD, sUSD, { value: twoETH, from: account1 }),
+					cerc20.open(oneRenBTC, toUnit(10000), sUSD, {
+						from: account1,
+					}),
 					'Exceed max borrow power'
 				);
 			});
 		});
-		describe('should open an eth loan denominated in sUSD', async () => {
+
+		describe('should open a btc loan denominated in sUSD', async () => {
+			const fiveHundredSUSD = toUnit(500);
+			let issueFeeRate;
+			let issueFee;
+
 			beforeEach(async () => {
-				tx = await ceth.open(fiveHundredsUSD, sUSD, {
-					value: tenETH,
+				tx = await cerc20.open(oneRenBTC, fiveHundredSUSD, sUSD, {
 					from: account1,
 				});
 
 				id = getid(tx);
 
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
+
+				issueFeeRate = new BN(await cerc20.issueFeeRate());
+				issueFee = fiveHundredSUSD.mul(issueFeeRate);
 			});
 
 			it('should set the loan correctly', async () => {
 				assert.equal(loan.account, account1);
-				assert.equal(loan.collateral, tenETH.toString());
+				assert.equal(loan.collateral, toUnit(1).toString());
 				assert.equal(loan.currency, sUSD);
-				assert.equal(loan.amount, fiveHundredsUSD.toString());
+				assert.equal(loan.amount, fiveHundredSUSD.toString());
 				assert.bnEqual(loan.accruedInterest, toUnit(0));
 			});
 
 			it('should issue the correct amount to the borrower', async () => {
-				// 0.001% issue fee rate.
-				const expectedBal = toUnit('499.5');
+				const expectedBal = fiveHundredSUSD.sub(issueFee);
 
 				assert.bnEqual(await sUSDSynth.balanceOf(account1), expectedBal);
 			});
@@ -374,60 +502,64 @@ contract('CollateralEth', async accounts => {
 			it('should issue the minting fee to the fee pool', async () => {
 				const feePoolBalance = await sUSDSynth.balanceOf(FEE_ADDRESS);
 
-				assert.bnEqual(toUnit('0.5'), feePoolBalance);
+				assert.equal(issueFee.toString(), feePoolBalance.toString());
 			});
 
 			it('should emit the event properly', async () => {
 				assert.eventEqual(tx, 'LoanCreated', {
 					account: account1,
 					id: id,
-					amount: fiveHundredsUSD,
-					collateral: tenETH,
+					amount: fiveHundredSUSD,
+					collateral: toUnit(1),
 					currency: sUSD,
 				});
 			});
 		});
 
-		describe('should open an eth loan denominated in sETH', async () => {
+		describe('should open a btc loan denominated in sBTC', async () => {
+			let issueFeeRate;
+			let issueFee;
+
 			beforeEach(async () => {
-				tx = await ceth.open(fiveETH, sETH, {
-					value: tenETH,
+				tx = await cerc20.open(fiveRenBTC, toUnit(2), sBTC, {
 					from: account1,
 				});
 
 				id = getid(tx);
 
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
+
+				issueFeeRate = await cerc20.issueFeeRate();
+				issueFee = toUnit(2).mul(issueFeeRate);
 			});
 
 			it('should set the loan correctly', async () => {
 				assert.equal(loan.account, account1);
-				assert.equal(loan.collateral, tenETH.toString());
-				assert.equal(loan.currency, sETH);
-				assert.equal(loan.amount, fiveETH.toString());
+				assert.equal(loan.collateral, toUnit(5).toString());
+				assert.equal(loan.currency, sBTC);
+				assert.equal(loan.amount, toUnit(2).toString());
 				assert.bnEqual(loan.accruedInterest, toUnit(0));
 			});
 
 			it('should issue the correct amount to the borrower', async () => {
-				// 0.001% issue fee rate.
-				const expectedBal = toUnit('4.995');
+				const expecetdBalance = toUnit(2).sub(issueFee);
 
-				assert.bnEqual(await sETHSynth.balanceOf(account1), expectedBal);
+				assert.bnEqual(await sBTCSynth.balanceOf(account1), expecetdBalance);
 			});
 
 			it('should issue the minting fee to the fee pool', async () => {
 				const feePoolBalance = await sUSDSynth.balanceOf(FEE_ADDRESS);
-				// usd equivalent of 0.005 ETH @ $100 per ETH.
-				assert.bnEqual(toUnit('0.5'), feePoolBalance);
+
+				assert.equal(issueFee.toString(), feePoolBalance.toString());
 			});
 
 			it('should emit the event properly', async () => {
 				assert.eventEqual(tx, 'LoanCreated', {
 					account: account1,
 					id: id,
-					amount: fiveETH,
-					collateral: tenETH,
-					currency: sETH,
+					amount: toUnit(2),
+					collateral: toUnit(5),
+					currency: sBTC,
 				});
 			});
 		});
@@ -435,8 +567,7 @@ contract('CollateralEth', async accounts => {
 
 	describe('deposits', async () => {
 		beforeEach(async () => {
-			tx = await ceth.open(100, sUSD, {
-				value: tenETH,
+			tx = await cerc20.open(twoRenBTC, oneHundredsUSD, sUSD, {
 				from: account1,
 			});
 
@@ -451,9 +582,9 @@ contract('CollateralEth', async accounts => {
 					beforeEach(async () => {
 						await setStatus({ owner, systemStatus, section, suspend: true });
 					});
-					it('then calling deposit() reverts', async () => {
+					it('then calling depopsit() reverts', async () => {
 						await assert.revert(
-							ceth.deposit(account1, id, { value: tenETH, from: account1 }),
+							cerc20.deposit(account1, id, oneRenBTC, { from: account1 }),
 							'Operation prohibited'
 						);
 					});
@@ -462,7 +593,7 @@ contract('CollateralEth', async accounts => {
 							await setStatus({ owner, systemStatus, section, suspend: false });
 						});
 						it('then calling deposit() succeeds', async () => {
-							await ceth.deposit(account1, id, { value: tenETH, from: account1 });
+							await cerc20.deposit(account1, id, oneRenBTC, { from: account1 });
 						});
 					});
 				});
@@ -472,7 +603,7 @@ contract('CollateralEth', async accounts => {
 		describe('revert conditions', async () => {
 			it('should revert if they do not send any eth', async () => {
 				await assert.revert(
-					ceth.deposit(account1, id, { value: 0, from: account1 }),
+					cerc20.deposit(account1, id, 0, { from: account1 }),
 					'Deposit must be above 0'
 				);
 			});
@@ -480,21 +611,22 @@ contract('CollateralEth', async accounts => {
 
 		describe('should allow deposits', async () => {
 			beforeEach(async () => {
-				await ceth.deposit(account1, id, { value: tenETH, from: account1 });
+				await cerc20.deposit(account1, id, oneRenBTC, { from: account1 });
 			});
 
 			it('should increase the total collateral of the loan', async () => {
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
 
-				assert.bnEqual(loan.collateral, twentyETH);
+				assert.bnEqual(loan.collateral, toUnit(3));
 			});
 		});
 	});
 
 	describe('withdraws', async () => {
+		let accountRenBalBefore;
+
 		beforeEach(async () => {
-			loan = await ceth.open(oneHundredsUSD, sUSD, {
-				value: tenETH,
+			loan = await cerc20.open(twoRenBTC, oneHundredsUSD, sUSD, {
 				from: account1,
 			});
 
@@ -509,9 +641,9 @@ contract('CollateralEth', async accounts => {
 					beforeEach(async () => {
 						await setStatus({ owner, systemStatus, section, suspend: true });
 					});
-					it('then calling withdraw() reverts', async () => {
+					it('then calling depopsit() reverts', async () => {
 						await assert.revert(
-							ceth.withdraw(id, oneETH, { from: account1 }),
+							cerc20.withdraw(id, oneRenBTC, { from: account1 }),
 							'Operation prohibited'
 						);
 					});
@@ -519,8 +651,8 @@ contract('CollateralEth', async accounts => {
 						beforeEach(async () => {
 							await setStatus({ owner, systemStatus, section, suspend: false });
 						});
-						it('then calling withdraw() succeeds', async () => {
-							await ceth.withdraw(id, oneETH, { from: account1 });
+						it('then calling deposit() succeeds', async () => {
+							await cerc20.withdraw(id, oneRenBTC, { from: account1 });
 						});
 					});
 				});
@@ -529,48 +661,42 @@ contract('CollateralEth', async accounts => {
 
 		describe('revert conditions', async () => {
 			it('should revert if the withdraw would put them under minimum collateralisation', async () => {
-				const nineETH = toUnit(9);
-				await assert.revert(ceth.withdraw(id, nineETH, { from: account1 }), 'Cratio too low');
+				await assert.revert(cerc20.withdraw(id, twoRenBTC, { from: account1 }), 'Cratio too low');
 			});
 
 			it('should revert if they try to withdraw all the collateral', async () => {
-				await assert.revert(ceth.withdraw(id, tenETH, { from: account1 }), 'Cratio too low');
+				await assert.revert(cerc20.withdraw(id, twoRenBTC, { from: account1 }), 'Cratio too low');
 			});
 
 			it('should revert if the sender is not borrower', async () => {
-				await assert.revert(ceth.withdraw(id, tenETH, { from: account2 }));
+				await issuesBTCtoAccount(oneRenBTC, account2);
+				await renBTC.approve(cerc20.address, oneRenBTC, { from: account2 });
+
+				await assert.revert(cerc20.withdraw(id, oneRenBTC, { from: account2 }));
 			});
 		});
 
 		describe('should allow withdraws', async () => {
 			beforeEach(async () => {
-				await ceth.withdraw(id, oneETH, {
+				accountRenBalBefore = await renBTC.balanceOf(account1);
+
+				await cerc20.withdraw(id, oneRenBTC, {
 					from: account1,
 				});
 			});
 
 			it('should decrease the total collateral of the loan', async () => {
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
 
-				const expectedCollateral = tenETH.sub(oneETH);
+				const expectedCollateral = toUnit(2).sub(toUnit(1));
 
 				assert.bnEqual(loan.collateral, expectedCollateral);
 			});
 
-			it('should create a pending withdraw entry', async () => {
-				const withdraw = await ceth.pendingWithdrawals(account1);
+			it('should transfer the withdrawn collateral to the borrower', async () => {
+				const bal = await renBTC.balanceOf(account1);
 
-				assert.bnEqual(withdraw, oneETH);
-			});
-
-			it('should allow the withdrawer to withdraw', async () => {
-				const bal = new BN(await getEthBalance(account1));
-
-				await ceth.claim(oneETH, { from: account1 });
-
-				const balAfter = new BN(await getEthBalance(account1));
-
-				assert.bnGt(balAfter, bal);
+				assert.bnEqual(bal, accountRenBalBefore.add(oneRenBTC));
 			});
 		});
 	});
@@ -578,8 +704,7 @@ contract('CollateralEth', async accounts => {
 	describe('repayments', async () => {
 		beforeEach(async () => {
 			// make a loan here so we have a valid ID to pass to the blockers and reverts.
-			tx = await ceth.open(oneHundredsUSD, sUSD, {
-				value: tenETH,
+			tx = await cerc20.open(twoRenBTC, oneHundredsUSD, sUSD, {
 				from: account1,
 			});
 
@@ -597,7 +722,7 @@ contract('CollateralEth', async accounts => {
 					});
 					it('then calling repay() reverts', async () => {
 						await assert.revert(
-							ceth.repay(account1, id, onesUSD, { from: account1 }),
+							cerc20.repay(account1, id, onesUSD, { from: account1 }),
 							'Operation prohibited'
 						);
 					});
@@ -606,7 +731,7 @@ contract('CollateralEth', async accounts => {
 							await setStatus({ owner, systemStatus, section, suspend: false });
 						});
 						it('then calling repay() succeeds', async () => {
-							await ceth.repay(account1, id, onesUSD, { from: account1 });
+							await cerc20.repay(account1, id, onesUSD, { from: account1 });
 						});
 					});
 				});
@@ -616,7 +741,7 @@ contract('CollateralEth', async accounts => {
 		describe('revert conditions', async () => {
 			it('should revert if they try to repay 0', async () => {
 				await assert.revert(
-					ceth.repay(account1, id, 0, { from: account1 }),
+					cerc20.repay(account1, id, 0, { from: account1 }),
 					'Payment must be above 0'
 				);
 			});
@@ -624,7 +749,7 @@ contract('CollateralEth', async accounts => {
 			// account 2 had no sUSD
 			it('should revert if they have no sUSD', async () => {
 				await assert.revert(
-					ceth.repay(account1, id, tensUSD, { from: account2 }),
+					cerc20.repay(account1, id, tensUSD, { from: account2 }),
 					'Not enough balance'
 				);
 			});
@@ -632,7 +757,7 @@ contract('CollateralEth', async accounts => {
 			it('should revert if they try to pay more than the amount owing', async () => {
 				await issuesUSDToAccount(toUnit(1000), account1);
 				await assert.revert(
-					ceth.repay(account1, id, toUnit(1000), { from: account1 }),
+					cerc20.repay(account1, id, toUnit(1000), { from: account1 }),
 					"VM Exception while processing transaction: reverted with reason string 'SafeMath: subtraction overflow'"
 				);
 			});
@@ -644,8 +769,8 @@ contract('CollateralEth', async accounts => {
 
 			beforeEach(async () => {
 				await issuesUSDToAccount(oneHundredsUSD, account2);
-				tx = await ceth.repay(account1, id, tensUSD, { from: account2 });
-				loan = await ceth.loans(id);
+				tx = await cerc20.repay(account1, id, tensUSD, { from: account2 });
+				loan = await cerc20.loans(id);
 			});
 
 			it('should work reduce the repayers balance', async () => {
@@ -654,27 +779,25 @@ contract('CollateralEth', async accounts => {
 			});
 
 			it('should update the loan', async () => {
-				assert.equal(loan.amount.toString().substring(0, 5), expectedString);
+				assert.bnClose(loan.amount.toString().substring(0, 5), expectedString);
 			});
 
-			it('should emit the event properly', async () => {
+			xit('should emit the event properly', async () => {
 				assert.eventEqual(tx, 'LoanRepaymentMade', {
 					account: account1,
 					repayer: account2,
 					id: id,
 					amountRepaid: tensUSD,
-					amountAfter: loan.amount,
+					amountAfter: parseInt(loan.amount),
 				});
 			});
 		});
 
-		describe('it should allow repayments on an sETH loan', async () => {
-			// I don't want to test interest here. I just want to test repayment.
-			const expectedString = '40000';
+		describe('it should allow repayments on an sBTC loan', async () => {
+			const expectedString = '10000';
 
 			beforeEach(async () => {
-				tx = await ceth.open(fiveETH, sETH, {
-					value: tenETH,
+				tx = await cerc20.open(fiveRenBTC, twoRenBTC, sBTC, {
 					from: account1,
 				});
 
@@ -682,48 +805,44 @@ contract('CollateralEth', async accounts => {
 
 				id = getid(tx);
 
-				await issuesETHToAccount(twoETH, account2);
+				await issuesBTCtoAccount(twoRenBTC, account2);
 
-				tx = await ceth.repay(account1, id, oneETH, { from: account2 });
+				tx = await cerc20.repay(account1, id, oneRenBTC, { from: account2 });
 
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
 			});
 
 			it('should work reduce the repayers balance', async () => {
-				const expectedBalance = oneETH;
+				const expectedBalance = oneRenBTC;
 
-				assert.bnEqual(await sETHSynth.balanceOf(account2), expectedBalance);
+				assert.bnEqual(await sBTCSynth.balanceOf(account2), expectedBalance);
 			});
 
 			it('should update the loan', async () => {
 				assert.equal(loan.amount.toString().substring(0, 5), expectedString);
 			});
 
-			it('should emit the event properly', async () => {
+			xit('should emit the event properly', async () => {
 				assert.eventEqual(tx, 'LoanRepaymentMade', {
 					account: account1,
 					repayer: account2,
 					id: id,
-					amountRepaid: oneETH,
-					amountAfter: loan.amount,
+					amountRepaid: oneRenBTC,
 				});
 			});
 		});
 	});
 
 	describe('liquidations', async () => {
-		let liquidatorEthBalBefore;
-
 		beforeEach(async () => {
 			// make a loan here so we have a valid ID to pass to the blockers and reverts.
-			loan = await ceth.open(toUnit('200'), sUSD, {
-				value: toUnit('2.6'),
+			tx = await cerc20.open(oneRenBTC, toUnit(5000), sUSD, {
 				from: account1,
 			});
 
 			await fastForwardAndUpdateRates(INTERACTION_DELAY);
 
-			id = getid(loan);
+			id = getid(tx);
 		});
 
 		describe('potential blocking conditions', async () => {
@@ -734,7 +853,7 @@ contract('CollateralEth', async accounts => {
 					});
 					it('then calling repay() reverts', async () => {
 						await assert.revert(
-							ceth.liquidate(account1, id, onesUSD, { from: account1 }),
+							cerc20.liquidate(account1, id, onesUSD, { from: account1 }),
 							'Operation prohibited'
 						);
 					});
@@ -745,7 +864,7 @@ contract('CollateralEth', async accounts => {
 						it('then calling liquidate() succeeds', async () => {
 							// fast forward a long time to make sure the loan is underwater.
 							await fastForwardAndUpdateRates(10 * YEAR);
-							await ceth.liquidate(account1, id, oneETH, { from: account1 });
+							await cerc20.liquidate(account1, id, onesUSD, { from: account1 });
 						});
 					});
 				});
@@ -755,36 +874,34 @@ contract('CollateralEth', async accounts => {
 		describe('revert conditions', async () => {
 			it('should revert if they have no sUSD', async () => {
 				await assert.revert(
-					ceth.liquidate(account1, id, onesUSD, { from: account2 }),
+					cerc20.liquidate(account1, id, onesUSD, { from: account2 }),
 					'Not enough balance'
 				);
 			});
 
 			it('should revert if they are not under collateralised', async () => {
 				await issuesUSDToAccount(toUnit(100), account2);
-				await ceth.deposit(account1, id, { value: oneETH, from: account1 });
-				await fastForwardAndUpdateRates(INTERACTION_DELAY);
 
 				await assert.revert(
-					ceth.liquidate(account1, id, onesUSD, { from: account2 }),
+					cerc20.liquidate(account1, id, onesUSD, { from: account2 }),
 					'Cratio above liq ratio'
 				);
 			});
 		});
 
 		describe('should allow liquidations on an undercollateralised sUSD loan', async () => {
-			const liquidatedCollateral = new BN('1588888888888888880');
+			const renAmount = new BN('19642857');
+			const internalAmount = new BN('196428571428571428');
 			let liquidationAmount;
 
 			beforeEach(async () => {
-				await updateAggregatorRates(exchangeRates, null, [sETH], [toUnit(90)]);
-				await issuesUSDToAccount(toUnit(1000), account2);
+				await updateAggregatorRates(exchangeRates, null, [sBTC], [7000].map(toUnit));
 
-				liquidatorEthBalBefore = new BN(await getEthBalance(account2));
+				await issuesUSDToAccount(toUnit(5000), account2);
 
-				liquidationAmount = await ceth.liquidationAmount(id);
+				liquidationAmount = await cerc20.liquidationAmount(id);
 
-				tx = await ceth.liquidate(account1, id, liquidationAmount, {
+				tx = await cerc20.liquidate(account1, id, liquidationAmount, {
 					from: account2,
 				});
 			});
@@ -795,20 +912,21 @@ contract('CollateralEth', async accounts => {
 					id: id,
 					liquidator: account2,
 					amountLiquidated: liquidationAmount,
-					collateralLiquidated: liquidatedCollateral,
+					collateralLiquidated: internalAmount,
 				});
 			});
 
 			it('should reduce the liquidators synth amount', async () => {
 				const liquidatorBalance = await sUSDSynth.balanceOf(account2);
-				const expectedBalance = toUnit('1000').sub(toUnit('130'));
-				assert.bnClose(liquidatorBalance, expectedBalance, '100000000000');
+				const expectedBalance = toUnit(5000).sub(liquidationAmount);
+
+				assert.bnEqual(liquidatorBalance, expectedBalance);
 			});
 
-			it('should create a pending withdrawl entry', async () => {
-				const withdaw = await ceth.pendingWithdrawals(account2);
+			it('should transfer the liquidated collateral to the liquidator', async () => {
+				const bal = await renBTC.balanceOf(account2);
 
-				assert.bnEqual(withdaw, liquidatedCollateral);
+				assert.bnEqual(bal, renAmount);
 			});
 
 			it('should pay the interest to the fee pool', async () => {
@@ -818,32 +936,22 @@ contract('CollateralEth', async accounts => {
 			});
 
 			it('should fix the collateralisation ratio of the loan', async () => {
-				const ratio = await ceth.collateralRatio(id);
+				const ratio = await cerc20.collateralRatio(id);
+
 				// the loan is very close 150%, we are in 10^18 land.
-				assert.bnClose(ratio, toUnit('1.3'), '10000000000000');
-			});
-
-			it('should allow the liquidator to call claim', async () => {
-				tx = await ceth.claim(liquidatedCollateral, { from: account2 });
-
-				const bal = new BN(await getEthBalance(account2));
-
-				assert.bnGt(bal, liquidatorEthBalBefore);
+				assert.bnClose(ratio, toUnit(1.5), '1500000000000');
 			});
 		});
 
 		describe('when a loan needs to be completely liquidated', async () => {
-			let liquidatorEthBalBefore;
-
 			beforeEach(async () => {
-				await updateAggregatorRates(exchangeRates, null, [sETH], [toUnit(50)]);
-				loan = await ceth.loans(id);
+				await updateAggregatorRates(exchangeRates, null, [sBTC], [5000].map(toUnit));
 
-				await issuesUSDToAccount(toUnit(1000), account2);
+				loan = await cerc20.loans(id);
 
-				liquidatorEthBalBefore = new BN(await getEthBalance(account2));
+				await issuesUSDToAccount(toUnit(10000), account2);
 
-				tx = await ceth.liquidate(account1, id, toUnit(1000), {
+				tx = await cerc20.liquidate(account1, id, toUnit(10000), {
 					from: account2,
 				});
 			});
@@ -854,49 +962,45 @@ contract('CollateralEth', async accounts => {
 					id: id,
 					liquidator: account2,
 					amountLiquidated: loan.amount,
-					collateralLiquidated: toUnit('2.6'),
+					collateralLiquidated: toUnit(1),
 				});
 			});
 
 			it('should close the loan correctly', async () => {
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
 
 				assert.equal(loan.amount, 0);
 				assert.equal(loan.collateral, 0);
 				assert.equal(loan.interestIndex, 0);
 			});
 
-			it('should reduce the liquidators synth amount', async () => {
-				const liquidatorBalance = await sUSDSynth.balanceOf(account2);
-				const expectedBalance = toUnit(1000).sub(toUnit('200'));
+			it('should transfer all the collateral to the liquidator', async () => {
+				const bal = await renBTC.balanceOf(account2);
 
-				assert.bnClose(liquidatorBalance, expectedBalance, '1000000000000000');
-			});
-
-			it('should create a pending withdrawl entry', async () => {
-				const withdaw = await ceth.pendingWithdrawals(account2);
-
-				assert.bnEqual(withdaw, toUnit('2.6'));
+				assert.bnEqual(bal, oneRenBTC);
 			});
 
 			it('should reduce the liquidators synth balance', async () => {
-				tx = await ceth.claim(twoETH, { from: account2 });
+				const liquidatorBalance = await sUSDSynth.balanceOf(account2);
+				const expectedBalance = toUnit(10000).sub(toUnit(5000));
 
-				const bal = new BN(await getEthBalance(account2));
-
-				assert.bnGt(bal, liquidatorEthBalBefore);
+				assert.bnClose(liquidatorBalance, expectedBalance, '10000000000000000');
 			});
 		});
 	});
 
 	describe('closing', async () => {
+		let accountRenBalBefore;
+
 		beforeEach(async () => {
+			accountRenBalBefore = await renBTC.balanceOf(account1);
+
 			// make a loan here so we have a valid ID to pass to the blockers and reverts.
-			loan = await ceth.open(oneHundredsUSD, sUSD, {
-				value: twoETH,
+			tx = await cerc20.open(twoRenBTC, oneHundredsUSD, sUSD, {
 				from: account1,
 			});
-			id = getid(loan);
+
+			id = getid(tx);
 
 			await fastForwardAndUpdateRates(INTERACTION_DELAY);
 		});
@@ -908,7 +1012,7 @@ contract('CollateralEth', async accounts => {
 						await setStatus({ owner, systemStatus, section, suspend: true });
 					});
 					it('then calling close() reverts', async () => {
-						await assert.revert(ceth.close(id, { from: account1 }), 'Operation prohibited');
+						await assert.revert(cerc20.close(id, { from: account1 }), 'Operation prohibited');
 					});
 					describe(`when ${section} is resumed`, () => {
 						beforeEach(async () => {
@@ -917,7 +1021,7 @@ contract('CollateralEth', async accounts => {
 						it('then calling close() succeeds', async () => {
 							// Give them some more sUSD to make up for the fees.
 							await issuesUSDToAccount(tensUSD, account1);
-							await ceth.close(id, { from: account1 });
+							await cerc20.close(id, { from: account1 });
 						});
 					});
 				});
@@ -926,11 +1030,11 @@ contract('CollateralEth', async accounts => {
 
 		describe('revert conditions', async () => {
 			it('should revert if they have no sUSD', async () => {
-				await assert.revert(ceth.close(id, { from: account1 }), 'Not enough balance');
+				await assert.revert(cerc20.close(id, { from: account1 }), 'Not enough balance');
 			});
 
 			it('should revert if they are not the borrower', async () => {
-				await assert.revert(ceth.close(id, { from: account2 }), 'Must be borrower');
+				await assert.revert(cerc20.close(id, { from: account2 }), 'Must be borrower');
 			});
 		});
 
@@ -939,11 +1043,11 @@ contract('CollateralEth', async accounts => {
 				// Give them some more sUSD to make up for the fees.
 				await issuesUSDToAccount(tensUSD, account1);
 
-				tx = await ceth.close(id, { from: account1 });
+				tx = await cerc20.close(id, { from: account1 });
 			});
 
 			it('should record the loan as closed', async () => {
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
 
 				assert.equal(loan.amount, 0);
 				assert.equal(loan.collateral, 0);
@@ -957,18 +1061,9 @@ contract('CollateralEth', async accounts => {
 				assert.bnGt(balance, toUnit(0));
 			});
 
-			it('should add a pending withdrawl entry', async () => {
-				assert.bnEqual(await ceth.pendingWithdrawals(account1), twoETH);
-			});
-
-			it('should allow the closer to withdraw', async () => {
-				const bal = new BN(await getEthBalance(account1));
-
-				await ceth.claim(oneETH, { from: account1 });
-
-				const balAfter = new BN(await getEthBalance(account1));
-
-				assert.bnGt(balAfter, bal);
+			it('should transfer the collateral back to the borrower', async () => {
+				const bal = await renBTC.balanceOf(account1);
+				assert.bnEqual(bal, accountRenBalBefore);
 			});
 
 			it('should emit the event', async () => {
@@ -983,8 +1078,7 @@ contract('CollateralEth', async accounts => {
 	describe('drawing', async () => {
 		beforeEach(async () => {
 			// make a loan here so we have a valid ID to pass to the blockers and reverts.
-			tx = await ceth.open(oneHundredsUSD, sUSD, {
-				value: twoETH,
+			tx = await cerc20.open(oneRenBTC, fiveThousandsUSD, sUSD, {
 				from: account1,
 			});
 
@@ -1000,14 +1094,17 @@ contract('CollateralEth', async accounts => {
 						await setStatus({ owner, systemStatus, section, suspend: true });
 					});
 					it('then calling draw() reverts', async () => {
-						await assert.revert(ceth.draw(id, onesUSD, { from: account1 }), 'Operation prohibited');
+						await assert.revert(
+							cerc20.draw(id, onesUSD, { from: account1 }),
+							'Operation prohibited'
+						);
 					});
 					describe(`when ${section} is resumed`, () => {
 						beforeEach(async () => {
 							await setStatus({ owner, systemStatus, section, suspend: false });
 						});
 						it('then calling draw() succeeds', async () => {
-							await ceth.draw(id, onesUSD, {
+							await cerc20.draw(id, onesUSD, {
 								from: account1,
 							});
 						});
@@ -1019,14 +1116,14 @@ contract('CollateralEth', async accounts => {
 					await fastForward((await exchangeRates.rateStalePeriod()).add(web3.utils.toBN('300')));
 				});
 				it('then calling draw() reverts', async () => {
-					await assert.revert(ceth.draw(id, onesUSD, { from: account1 }), 'Invalid rate');
+					await assert.revert(cerc20.draw(id, onesUSD, { from: account1 }), 'Invalid rate');
 				});
-				describe('when ETH gets a rate', () => {
+				describe('when BTC gets a rate', () => {
 					beforeEach(async () => {
 						await updateRatesWithDefaults();
 					});
 					it('then calling draw() succeeds', async () => {
-						await ceth.draw(id, onesUSD, { from: account1 });
+						await cerc20.draw(id, onesUSD, { from: account1 });
 					});
 				});
 			});
@@ -1034,82 +1131,22 @@ contract('CollateralEth', async accounts => {
 
 		describe('revert conditions', async () => {
 			it('should revert if the draw would under collateralise the loan', async () => {
-				await assert.revert(ceth.draw(id, oneHundredsUSD, { from: account1 }), 'Cratio too low');
+				await fastForwardAndUpdateRates(INTERACTION_DELAY);
+
+				await assert.revert(cerc20.draw(id, toUnit(3000), { from: account1 }), 'Cratio too low');
 			});
 		});
 
 		describe('should draw the loan down', async () => {
 			beforeEach(async () => {
-				tx = await ceth.draw(id, toUnit(30), { from: account1 });
+				tx = await cerc20.draw(id, oneThousandsUSD, { from: account1 });
 
-				loan = await ceth.loans(id);
+				loan = await cerc20.loans(id);
 			});
 
 			it('should update the amount on the loan', async () => {
-				assert.equal(loan.amount, toUnit(130).toString());
+				assert.equal(loan.amount, toUnit(6000).toString());
 			});
-		});
-	});
-
-	describe('Accrue Interest', async () => {
-		beforeEach(async () => {
-			// 0.005% / 31556926 (seconds in common year)
-			await manager.setBaseBorrowRate(158443823, { from: owner });
-		});
-
-		it('should correctly determine the interest on loans', async () => {
-			tx = await ceth.open(oneHundredsUSD, sUSD, {
-				value: twoETH,
-				from: account1,
-			});
-
-			id = getid(tx);
-
-			// after a year we should have accrued about 0.005% + (100/2100) = 0.05261904762
-
-			await fastForwardAndUpdateRates(YEAR);
-
-			// deposit some eth to trigger the interest accrual.
-
-			tx = await ceth.deposit(account1, id, { from: account1, value: oneETH });
-
-			loan = await ceth.loans(id);
-
-			let interest = Math.round(parseFloat(fromUnit(loan.accruedInterest)) * 10000) / 10000;
-
-			assert.equal(interest, 5.2619);
-
-			tx = await ceth.open(oneHundredsUSD, sUSD, {
-				value: twoETH,
-				from: account1,
-			});
-
-			const id2 = getid(tx);
-
-			// after a year we should have accrued about 0.005% + (200/2200) = 0.09590909091
-
-			await fastForwardAndUpdateRates(YEAR);
-
-			tx = await ceth.deposit(account1, id2, { from: account1, value: oneETH });
-
-			loan = await ceth.loans(id2);
-
-			interest = Math.round(parseFloat(fromUnit(loan.accruedInterest)) * 10000) / 10000;
-
-			assert.equal(interest, 9.5909);
-
-			// after two years we should have accrued (this math is rough)
-			// 0.005% + (100/2100) = 0.05261904762 +
-			// 0.005% + (200/2200) = 0.09590909091 +
-			//                     = 0.1485281385
-
-			tx = await ceth.deposit(account1, id, { from: account1, value: oneETH });
-
-			loan = await ceth.loans(id);
-
-			interest = Math.round(parseFloat(fromUnit(loan.accruedInterest)) * 10000) / 10000;
-
-			assert.equal(interest, 14.8528);
 		});
 	});
 });
