+++ test/legacy/EtherWrapper.js
@@ -2,7 +2,7 @@
 
 const { contract } = require('hardhat');
 
-const { assert, addSnapshotBeforeRestoreAfterEach } = require('./common');
+const { assert, addSnapshotBeforeRestoreAfterEach } = require('../contracts/common');
 
 const { toUnit, multiplyDecimal } = require('../utils')();
 
@@ -12,9 +12,9 @@ const {
 	decodedEventEqual,
 	setupPriceAggregators,
 	updateAggregatorRates,
-} = require('./helpers');
+} = require('../contracts/helpers');
 
-const { setupAllContracts } = require('./setup');
+const { setupAllContracts } = require('../contracts/setup');
 
 const { toBytes32 } = require('../..');
 const { toBN } = require('web3-utils');
