+++ test/legacy/VirtualSynthMastercopy.unit.js
@@ -2,9 +2,12 @@
 
 const { artifacts, contract } = require('hardhat');
 
-const { assert } = require('./common');
+const { assert } = require('../contracts/common');
 
-const { ensureOnlyExpectedMutativeFunctions, trimUtf8EscapeChars } = require('./helpers');
+const {
+	ensureOnlyExpectedMutativeFunctions,
+	trimUtf8EscapeChars,
+} = require('../contracts/helpers');
 
 const {
 	toBytes32,
